% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_binary_vectors}
\alias{all_binary_vectors}
\title{Create the list of all the possible binary vectors of given length, sorted lexicographically.}
\usage{
all_binary_vectors(n)
}
\arguments{
\item{n}{A length of the vectors.}
}
\value{
The list of the vectors.
}
\description{
Create the list of all the possible binary vectors of given length, sorted lexicographically.
}
