% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{bifurcation_plot}
\alias{bifurcation_plot}
\title{Draws the bifurcation plot for the given data}
\usage{
bifurcation_plot(discrete_bifurcation_result, main = "")
}
\arguments{
\item{discrete_bifurcation_result}{the quantitative discrete bifurcation returned by compute_discrete_bifurcation}

\item{main}{the title for the plot}
}
\value{
The bifurcation plot
}
\description{
Draws the bifurcation plot for the given data
}
