% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binary_vector_to_number}
\alias{binary_vector_to_number}
\title{Get the number represented by binary vector (from most to least significant bit).}
\usage{
binary_vector_to_number(binary_vector)
}
\description{
Get the number represented by binary vector (from most to least significant bit).
}
