% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binary_vector_to_one_based_index}
\alias{binary_vector_to_one_based_index}
\title{Get the 1-based index of a binary vector representing a row in a truth-table.}
\usage{
binary_vector_to_one_based_index(binary_vector)
}
\description{
Get the 1-based index of a binary vector representing a row in a truth-table.
}
