% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities.R
\name{compute_activity_ratios_matrix}
\alias{compute_activity_ratios_matrix}
\title{Computes matrix of activity ratios (output genes x attractors)}
\usage{
compute_activity_ratios_matrix(
  attractors,
  genes,
  output_genes,
  output_genes_encoder,
  instance_states
)
}
\description{
Computes matrix of activity ratios (output genes x attractors)
}
