% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attractors.R
\name{compute_attractor_distribution_matrix}
\alias{compute_attractor_distribution_matrix}
\title{Compute the async attractor distributions of all states in a form of matrix.}
\usage{
compute_attractor_distribution_matrix(
  instance,
  instance_states,
  attractors,
  attractor_membership_vector
)
}
\description{
Compute the async attractor distributions of all states in a form of matrix.
}
