% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities.R
\name{compute_attractor_distribution_similarity}
\alias{compute_attractor_distribution_similarity}
\title{Computes the attractor distribution similarity using LP}
\usage{
compute_attractor_distribution_similarity(
  attractor_distribution_1,
  attractor_distribution_2,
  lp_objective_coefficients,
  lp_constraints_coefficients_matrix,
  lp_directions
)
}
\description{
Computes the attractor distribution similarity using LP
}
