% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities.R
\name{compute_attractor_landscape_similarity}
\alias{compute_attractor_landscape_similarity}
\title{Computes attractor landscape similarity for two given parametrisations}
\usage{
compute_attractor_landscape_similarity(
  attractor_landscape_1,
  attractor_landscape_2,
  genes,
  output_genes,
  output_genes_encoder,
  instance_states,
  attractor_similarity = "activity"
)
}
\description{
Computes attractor landscape similarity for two given parametrisations
}
