% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attractors.R
\name{compute_attractor_membership_vector}
\alias{compute_attractor_membership_vector}
\title{Get the vector of states membership in attractors.}
\usage{
compute_attractor_membership_vector(instance, instance_states, attractors)
}
\description{
Get the vector of states membership in attractors.
}
