% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities.R
\name{compute_attractors_output_states_ratios_matrix}
\alias{compute_attractors_output_states_ratios_matrix}
\title{Computes matrix of output states ratios (output states x attractors)}
\usage{
compute_attractors_output_states_ratios_matrix(
  attractors,
  output_genes,
  output_genes_encoder,
  instance_states
)
}
\description{
Computes matrix of output states ratios (output states x attractors)
}
