% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities.R
\name{compute_attractors_similarity_matrix}
\alias{compute_attractors_similarity_matrix}
\title{Computes matrix of similarities for two given sets of attractors, using given attractor similarity metric}
\usage{
compute_attractors_similarity_matrix(
  attractors_1,
  attractors_2,
  genes,
  output_genes,
  output_genes_encoder,
  instance_states,
  attractor_similarity
)
}
\description{
Computes matrix of similarities for two given sets of attractors, using given attractor similarity metric
}
