% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities.R
\name{compute_attractors_similarity_matrix_activity}
\alias{compute_attractors_similarity_matrix_activity}
\title{Computes matrix of activity similarities for two given sets of attractors}
\usage{
compute_attractors_similarity_matrix_activity(
  attractors_1,
  attractors_2,
  genes,
  output_genes,
  output_genes_encoder,
  instance_states
)
}
\description{
Computes matrix of activity similarities for two given sets of attractors
}
