% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities.R
\name{compute_attractors_similarity_matrix_overlap}
\alias{compute_attractors_similarity_matrix_overlap}
\title{Computes matrix of overlap similarities for two given sets of attractors}
\usage{
compute_attractors_similarity_matrix_overlap(
  attractors_1,
  attractors_2,
  output_genes,
  output_genes_encoder,
  instance_states
)
}
\description{
Computes matrix of overlap similarities for two given sets of attractors
}
