% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities.R
\name{compute_constraints_coefficients_matrix}
\alias{compute_constraints_coefficients_matrix}
\title{Computes the matrix of constrain coefficients for LP computation}
\usage{
compute_constraints_coefficients_matrix(attractor_1_count, attractor_2_count)
}
\description{
Computes the matrix of constrain coefficients for LP computation
}
