% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance.R
\name{compute_conversion_vector}
\alias{compute_conversion_vector}
\title{Compute the conversion vector for input genes of a given gene.
The resulting vector contains indexes of input genes with regards to the all_genes vector.}
\usage{
compute_conversion_vector(instance, all_genes, gene)
}
\description{
Compute the conversion vector for input genes of a given gene.
The resulting vector contains indexes of input genes with regards to the all_genes vector.
}
