% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{compute_db_single}
\alias{compute_db_single}
\title{The QDBF copmutation for one given parametrisation}
\usage{
compute_db_single(
  index_combination_vector,
  gene_function_vectors,
  baseline_parametrisation,
  baseline_attractor_landscape,
  output_genes,
  output_genes_encoder,
  genes,
  gene_regulators_indexes,
  instance_states,
  pre_to_post,
  semantics,
  attractor_similarity
)
}
\description{
The QDBF copmutation for one given parametrisation
}
