% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{compute_discrete_bifurcation}
\alias{compute_discrete_bifurcation}
\title{Computes the quantitative discrete bifurcation function of the
given baseline BN in the parametrisation space represented by the given RG.
If necessary, the motif extension is created before computation.}
\usage{
compute_discrete_bifurcation(
  bn_baseline_network,
  rg_rewiring_space,
  output_genes = character(),
  semantics = "async",
  attractor_similarity = "activity",
  return_pbn = FALSE
)
}
\arguments{
\item{bn_baseline_network}{the baseline BN. Either a file path or a BooleanNetworkin-stance returned by load_BN.}

\item{rg_rewiring_space}{the RG representing the parametrisation space. Either a file path or an RG data frame.}

\item{output_genes}{the vector of output genes. Must be a subset of genes inbn. If not provided, defaults to BN genes}

\item{semantics}{the BN semantics to be used:"async"or"sync"}

\item{attractor_similarity}{the attractor similarity measure to be used:"activity"or"overlap"}

\item{return_pbn}{the logical value indicating whether to also return the constructed PBN}
}
\value{
The table of rewiring distance and attractor landscape similarity for all parametrisations.
}
\description{
Computes the quantitative discrete bifurcation function of the
given baseline BN in the parametrisation space represented by the given RG.
If necessary, the motif extension is created before computation.
}
