% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{compute_discrete_bifurcation_inner}
\alias{compute_discrete_bifurcation_inner}
\title{The inner computation of QDBF.}
\usage{
compute_discrete_bifurcation_inner(
  pbn,
  baseline_parametrisation,
  output_genes = character(),
  semantics = "async",
  attractor_similarity = "activity"
)
}
\description{
The inner computation of QDBF.
}
