% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\name{compute_intermediate_genes}
\alias{compute_intermediate_genes}
\title{Computes the set of transitive regulators of motif genes (including them).}
\usage{
compute_intermediate_genes(motif_genes, rg_large)
}
\arguments{
\item{motif_genes}{the genes of the motif}

\item{rg_large}{the RG to find the transitive closure in}
}
\value{
vector of all the transitive regulators
}
\description{
Computes the set of transitive regulators of motif genes (including them).
}
