% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities.R
\name{compute_rewiring_distance}
\alias{compute_rewiring_distance}
\title{Computes rewiring distance for two given parametrisations}
\usage{
compute_rewiring_distance(parametrisation_1, parametrisation_2)
}
\description{
Computes rewiring distance for two given parametrisations
}
