% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{compute_rewiring_robustness}
\alias{compute_rewiring_robustness}
\title{Computes the phenotype rewiring robustness over the data result-ing from a computation ofcompute_discrete_bifurcation.}
\usage{
compute_rewiring_robustness(discrete_bifurcation_result, p_elementary = 0.1)
}
\arguments{
\item{discrete_bifurcation_result}{the quantitative discrete bifurcation returned by compute_discrete_bifurcation}

\item{p_elementary}{the elementary rewiring probability}
}
\value{
The value of phenotype rewiring robustness
}
\description{
Computes the phenotype rewiring robustness over the data result-ing from a computation ofcompute_discrete_bifurcation.
}
