% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance.R
\name{compute_transition_table_cell}
\alias{compute_transition_table_cell}
\title{Compute value of one cell of transition table, given input row and column gene.}
\usage{
compute_transition_table_cell(
  instance,
  conversion_vectors,
  input_row,
  cell_gene
)
}
\description{
Compute value of one cell of transition table, given input row and column gene.
}
