% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance.R
\name{compute_transition_table_row}
\alias{compute_transition_table_row}
\title{Compute one row of transition table.}
\usage{
compute_transition_table_row(instance, conversion_vectors, input_row)
}
\description{
Compute one row of transition table.
}
