% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\name{create_extension}
\alias{create_extension}
\title{Transform the given motif BN and possibly larger PBN to an intermediate gene set.}
\usage{
create_extension(bn_motif, rg_large, output_genes)
}
\description{
Transform the given motif BN and possibly larger PBN to an intermediate gene set.
}
