% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\name{create_intermediate_motif_parametrisation}
\alias{create_intermediate_motif_parametrisation}
\title{Compute the parametrisation representing teh original motig in the extension RG}
\usage{
create_intermediate_motif_parametrisation(bn_motif, pbn_intermediate)
}
\description{
Compute the parametrisation representing teh original motig in the extension RG
}
