% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\name{create_intermediate_rg}
\alias{create_intermediate_rg}
\title{Compute the motif extension from the given RG over given genes}
\usage{
create_intermediate_rg(rg, intermediate_genes)
}
\description{
Compute the motif extension from the given RG over given genes
}
