% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attractors.R
\name{decode_attractor}
\alias{decode_attractor}
\title{Decode attractor from Boolnet into a vector of (binary) numbers representing states.}
\usage{
decode_attractor(attractor, pre_to_post)
}
\description{
Decode attractor from Boolnet into a vector of (binary) numbers representing states.
}
