% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbn.R
\name{enumerate_gene_function_vectors}
\alias{enumerate_gene_function_vectors}
\title{Enumerates all valid function vectors for given gene}
\usage{
enumerate_gene_function_vectors(rg, gene)
}
\description{
Enumerates all valid function vectors for given gene
}
