% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance.R
\name{extract_instance}
\alias{extract_instance}
\title{Extract a BN instance from a PBN for a given index.}
\usage{
extract_instance(pbn, instance_index)
}
\arguments{
\item{pbn}{The PBN to extract from}

\item{instance_index}{1-based numeric index of a parametrisation in the PBN}
}
\value{
the BN instance
}
\description{
Extract a BN instance from a PBN for a given index.
}
