% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbn.R
\name{generate_BN}
\alias{generate_BN}
\title{Generates a BooleanNetwork instance from the given data}
\usage{
generate_BN(genes, gene_regulators_indexes, gene_function_vectors)
}
\description{
Generates a BooleanNetwork instance from the given data
}
