% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbn.R
\name{generate_interaction}
\alias{generate_interaction}
\title{Generates a single interaction for genereated BN}
\usage{
generate_interaction(gene_i, gene_regulators_indexes, gene_function_vectors)
}
\description{
Generates a single interaction for genereated BN
}
