% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbn.R
\name{generate_parametrisation_instance}
\alias{generate_parametrisation_instance}
\title{Create a BN from a given parametrisation}
\usage{
generate_parametrisation_instance(pbn, parametrisation)
}
\description{
Create a BN from a given parametrisation
}
