% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{get_attractor_landscape}
\alias{get_attractor_landscape}
\title{Get the attractor landscape of a given parametrisation in the form of attractors list
and attractor distribution matrix.}
\usage{
get_attractor_landscape(
  parametrisation,
  genes,
  gene_regulators_indexes,
  instance_states,
  pre_to_post,
  semantics = "async"
)
}
\description{
Get the attractor landscape of a given parametrisation in the form of attractors list
and attractor distribution matrix.
}
