% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attractors.R
\name{get_attractors_data_async}
\alias{get_attractors_data_async}
\title{Get the list of attractors and the attractor probability distributions of the states under asynchronous semantics.}
\usage{
get_attractors_data_async(instance, instance_states, pre_to_post)
}
\description{
Get the list of attractors and the attractor probability distributions of the states under asynchronous semantics.
}
