% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attractors.R
\name{get_attractors_data_sync}
\alias{get_attractors_data_sync}
\title{Get the list of attractors and the attractor probability distributions of the states under synchronous semantics.}
\usage{
get_attractors_data_sync(instance, pre_to_post)
}
\description{
Get the list of attractors and the attractor probability distributions of the states under synchronous semantics.
}
