% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\name{get_function_vector}
\alias{get_function_vector}
\title{Get the function vector of given gene for the baseline parametrisation in motif extension}
\usage{
get_function_vector(gene_i, bn_motif, pbn_intermediate, motif_transition_table)
}
\description{
Get the function vector of given gene for the baseline parametrisation in motif extension
}
