% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbn.R
\name{get_parametrisation_by_index}
\alias{get_parametrisation_by_index}
\title{Get a parametrisation from PBN by its numeric 1-based index.}
\usage{
get_parametrisation_by_index(pbn, parametrisation_index)
}
\description{
Get a parametrisation from PBN by its numeric 1-based index.
}
