% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbn.R
\name{get_parametrisation_by_index_combination_vector}
\alias{get_parametrisation_by_index_combination_vector}
\title{Get a parametrisation from PBN by vector of indexes of update functions of respective genes.}
\usage{
get_parametrisation_by_index_combination_vector(
  gene_function_vectors,
  index_combination_vector
)
}
\description{
Get a parametrisation from PBN by vector of indexes of update functions of respective genes.
}
