% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbn.R
\name{is_function_vector_valid}
\alias{is_function_vector_valid}
\title{Validates that given function vector satisfies the constraints given by the regulatory graph.}
\usage{
is_function_vector_valid(
  function_vector,
  regulation_count,
  regulations,
  input_rows
)
}
\description{
Validates that given function vector satisfies the constraints given by the regulatory graph.
}
