% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_BN}
\alias{load_BN}
\title{Load the BoolNet Boolean network from the given .bn file.}
\usage{
load_BN(file_name)
}
\arguments{
\item{file_name}{File name to load from.}
}
\value{
The Boolean network.
}
\description{
Load the BoolNet Boolean network from the given .bn file.
}
