% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_PBN}
\alias{load_PBN}
\title{Load the parametrised Boolean network from the RG in given file.}
\usage{
load_PBN(file_name)
}
\arguments{
\item{file_name}{File name to load from.}
}
\value{
The parametrised Boolean network.
}
\description{
Load the parametrised Boolean network from the RG in given file.
}
