% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_RG}
\alias{load_RG}
\title{Load the Regulatory graph from the given file.}
\usage{
load_RG(file_name)
}
\arguments{
\item{file_name}{File name to load from.}
}
\value{
The Regulatory graph.
}
\description{
Load the Regulatory graph from the given file.
}
