% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelism.R
\name{parallel_apply}
\alias{parallel_apply}
\title{Apply a function to all members of a collection parallelly.}
\usage{
parallel_apply(
  collection,
  fun,
  packages = character(),
  seq_threshold = 0,
  log_frequency = 0
)
}
\arguments{
\item{collection}{the input collection, either a list or vector}

\item{fun}{function to apply}

\item{packages}{the string vector of all packages used in the function (necessary!)}

\item{seq_threshold}{the threshold, under which sequential execution should be chosen. If 0, it does not apply.}

\item{log_frequency}{How often (as per number of processed items in the collection) to log. 0 means no logging}
}
\value{
The result - a collection of results of function applications to the input collection, in the same order.
}
\description{
Apply a function to all members of a collection parallelly.
}
