% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pre_to_post_vector}
\alias{pre_to_post_vector}
\title{Create the pre-to-post vector for decoding attractor states returned by Boolnet.}
\usage{
pre_to_post_vector(states)
}
\description{
Create the pre-to-post vector for decoding attractor states returned by Boolnet.
}
