% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelism.R
\name{prepare_parallel}
\alias{prepare_parallel}
\title{Prepares the parallel execution. Creates a cluster for a number of actual CPU cores minus one.}
\usage{
prepare_parallel()
}
\value{
the cluster.
}
\description{
Prepares the parallel execution. Creates a cluster for a number of actual CPU cores minus one.
}
