% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{process_inputs}
\alias{process_inputs}
\title{Process the inputs given to compute_discrete_bifurcation}
\usage{
process_inputs(bn_baseline_network, rg_rewiring_space, output_genes)
}
\description{
Process the inputs given to compute_discrete_bifurcation
}
