% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\name{validate_edges}
\alias{validate_edges}
\title{Validate the proper relationship between the edges of motif and RG.}
\usage{
validate_edges(bn_motif, rg_large)
}
\description{
Validate the proper relationship between the edges of motif and RG.
}
