% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\name{validate_genes}
\alias{validate_genes}
\title{Validate the proper relationship between the genes of output, motif, and RG.}
\usage{
validate_genes(bn_motif, rg_large, output_genes)
}
\description{
Validate the proper relationship between the genes of output, motif, and RG.
}
