% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbn.R
\name{validate_unique_edges}
\alias{validate_unique_edges}
\title{Validate that there are no repeating edges in the RG}
\usage{
validate_unique_edges(rg)
}
\description{
Validate that there are no repeating edges in the RG
}
