% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\name{compute_intermediate_genes}
\alias{compute_intermediate_genes}
\title{Compute the set of transitive regulators of motif genes}
\usage{
compute_intermediate_genes(bn_motif, rg_large)
}
\description{
Compute the set of transitive regulators of motif genes
}
