% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension.R
\name{create_intermediate_rg}
\alias{create_intermediate_rg}
\title{Computes the motif extension from the given RG over given genes}
\usage{
create_intermediate_rg(rg, intermediate_genes)
}
\arguments{
\item{rg}{the large rg}

\item{intermediate_genes}{the set of genes to narrow to}
}
\value{
the graph narrowed to the given set of regulated genes
}
\description{
Computes the motif extension from the given RG over given genes
}
